Renamed default entity manager.
This patch includes a small fix for when the default entity manager
is being used and you attempt to get a fixture. The hasFixture()
method will return true if the fixture exists, but getFixture()
will go on to fail because the entity manager name is still null.

This fixes that by making two calls to getEntityManagerName() which
will return the name of the entity manager if it exists or __default__
otherwise.